refactor(domain): relocate password hashing and canonical email out of the test package
Production code (dynamodb-auth.ts, app.ts) imported the real hashPassword/verifyPassword from @packages/test-fixtures, making production depend on a test package. Move those credential primitives — and the canonicalizeEmail/CanonicalEmail identity key — into @packages/domain/user alongside normalizeEmail, so nothing production-bound resolves from @packages/test-fixtures (which now keeps only its in-memory test doubles).

Add a verifyPassword(undefined) case so password.ts stays at 100% coverage under domain's threshold.
