+++ projects/hutch/src/runtime/app.ts
@@ -3,7 +3,8 @@ import assert from "node:assert";
 import type { Express } from "express";
 import { createDynamoDocumentClient } from "@packages/hutch-storage-client";
 import type { Logger } from "./domain/logger";
-import { initInMemoryAuth, hashPassword, verifyPassword } from "@packages/test-fixtures/providers/auth";
+import { initInMemoryAuth } from "@packages/test-fixtures/providers/auth";
+import { hashPassword, verifyPassword } from "@packages/domain/user";
 import { initDynamoDbAuth } from "./providers/auth/dynamodb-auth";
 import { initInMemoryArticleStore } from "@packages/test-fixtures/providers/article-store";
 import { initDynamoDbArticleStore } from "./providers/article-store/dynamodb-article-store";
