+++ src/packages/domain/src/user/index.ts
@@ -1,5 +1,7 @@
 export type { UserId, AuthenticatedUserId } from "./user.types";
 export { UserIdSchema, authenticatedUserIdFrom } from "./user.schema";
-export { normalizeEmail } from "./email";
+export { normalizeEmail, canonicalizeEmail } from "./email";
+export type { CanonicalEmail } from "./email";
+export { hashPassword, verifyPassword } from "./password";
 export type { UserIdPrefix } from "./user-id-prefix";
 export { USER_ID_PREFIX_LENGTH, UserIdPrefixSchema, userIdPrefixFrom, parseUserIdPrefix } from "./user-id-prefix";
