--- src/packages/test-fixtures/src/providers/auth/canonicalize-email.ts
@@ -1,43 +1,0 @@
-import assert from "node:assert";
-import { z } from "zod";
-
-const CanonicalEmailSchema = z.string().brand<"CanonicalEmail">();
-
-/**
- * The identity key for an account — NOT a delivery address. `canonicalizeEmail`
- * collapses the many Gmail spellings of one mailbox (dots, +tags, googlemail.com)
- * onto a single key so they cannot become separate accounts. Mail must always be
- * sent to the address the user actually typed (see `normalizeEmail`), never to this key.
- */
-export type CanonicalEmail = z.infer<typeof CanonicalEmailSchema>;
-
-const GMAIL_DOMAINS = new Set(["gmail.com", "googlemail.com"]);
-
-const GMAIL_PLAINTEXT_BASE = /^[A-Za-z0-9.]+$/;
-
-export function canonicalizeEmail(raw: string): CanonicalEmail {
-	const trimmed = raw.trim();
-	const at = trimmed.lastIndexOf("@");
-	assert(at > 0, "Email is missing a local part");
-	assert(at < trimmed.length - 1, "Email is missing a domain");
-
-	const local = trimmed.slice(0, at);
-	const loweredDomain = trimmed.slice(at + 1).toLowerCase();
-	const domain = loweredDomain.endsWith(".") ? loweredDomain.slice(0, -1) : loweredDomain;
-
-	if (!GMAIL_DOMAINS.has(domain)) {
-		return CanonicalEmailSchema.parse(`${local}@${domain}`);
-	}
-
-	const plusIndex = local.indexOf("+");
-	const base = plusIndex === -1 ? local : local.slice(0, plusIndex);
-
-	if (base.length > 0 && !GMAIL_PLAINTEXT_BASE.test(base)) {
-		return CanonicalEmailSchema.parse(`${local}@gmail.com`);
-	}
-
-	const canonical = base.toLowerCase().replaceAll(".", "");
-	assert(canonical.length > 0, "Gmail address has no local part before the tag");
-
-	return CanonicalEmailSchema.parse(`${canonical}@gmail.com`);
-}
