+++ src/packages/test-fixtures/src/providers/auth/index.ts
@@ -1,6 +1,4 @@
 export * from "./auth.types";
 export * from "./bot-defense.types";
-export * from "./canonicalize-email";
 export * from "./conversion.types";
 export * from "./in-memory-auth";
-export * from "./password";
